rename .c and .h
